requirements.txt
boto3==1.26.74
botocore==1.29.74
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.7
cryptography==35.0.0
Deprecated==1.2.18
distlib==0.3.9
Django==5.2.4
django-cors-headers==4.7.0
django-rest-knox==5.0.2
django-storages==1.14.6
djangorestframework==3.16.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.9.0
flake8==6.0.0
pre-commit==2.21.0
psycopg2-binary==2.9.10
PyJWT==2.1.0
PyMySQL==1.1.1
pytest==7.2.0
pytest-asyncio==0.24.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.7.4
s3transfer==0.6.0
